refactor(ts): refactor pre-processor files
