--- src/proxy/processors/pre-processor/index.js
@@ -1,1 +1,0 @@
-exports.parseAction = require('./parseAction').exec;
