+++ src/proxy/processors/pre-processor/index.ts
@@ -1,0 +1,3 @@
+import { exec } from './parseAction';
+
+export const parseAction = exec;
